+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/recorder/Recorder.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.audiorecorder.recording
+package org.odk.collect.audiorecorder.recorder
 
 import java.io.File
