+++ audiorecorder/src/test/java/org/odk/collect/audiorecorder/recording/MediaRecorderRecorderTest.kt
@@ -6,6 +6,8 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.hamcrest.Matchers.not
 import org.junit.Test
+import org.odk.collect.audiorecorder.recorder.MediaRecorderRecorder
+import org.odk.collect.audiorecorder.recorder.MediaRecorderWrapper
 import java.io.File
 
 class MediaRecorderRecorderTest {
